requirements.txt
Django~=5.1.2
django-mptt~=0.16.0
django-debug-toolbar~=4.4.6
pillow~=11.0.0
django-versatileimagefield~=3.1
python-magic-bin~=0.4.14